+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2022-12-14
+
+### Added
+
+* `dependencies create` command: More detailed explanation of the "Dependencies" template.
+* `dependencies create` command: `--quiet` option
 
 ## [0.2.1] - 2022-12-14